requirements.txt
click
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
htmlmin
jinja2
requests
